[gocui] Fix attributes of inactive tab titles if activeBorderColor doesn't include bold
The code was written with the assumption that the active border color includes
the bold attribute, which is true for the default theme color; however, when
changing the scheme to a color that doesn't include bold, it would result in
many wrong attribute bits being set.

Fix this by using bitwise operations instead of arithmetic operations to clear
the bold bit.